requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.7
pinecone==7.3.0
langgraph==0.0.26
langchain==0.0.345
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-core==0.1.0
requests==2.31.0
aiofiles==23.2.1
